==> data-raw/00-Jacksonville-Florida-Basic-Data.Rmd <==
---
title: "Jacksonville Florida"
---

# Metro Data

This will be the basic data pre-processing work for the metro region. The other city notebooks follow the same steps but with different sets of CBSA data.

Estimated time to run: a few minutes

Load packages:

```{r}
library(ggplot2)
library(glue) # Interpreted String Literals
#library(h3)
library(here) # A Simpler Way to Find Your Files
#library(osmdata) # package for working with streets
library(sf) # Simple Features for R
library(tidycensus) # Load US Census Boundary and Attribute Data as 'tidyverse' and 'sf'-Ready Data Frames
library(tidyverse) # Easily Install and Load the 'Tidyverse'
library(tigris) # Load Census TIGER/Line Shapefiles
```

## Census tracts

Obtain the census tracts for the metro of interest. Note that [{tidycensus}](https://walker-data.com/tidycensus/) function `get_decenial()` allows the selection of state and county, but not metro area or city. Check the example [here](https://walker-data.com/census-r/spatial-analysis-with-us-census-data.html#identifying-geometries-within-a-metropolitan-area) to see how to select tracts in a city.

```{r}
options(tigris_use_cache = TRUE)

fl_tracts <- get_decennial(geography = "tract",
                           variables = "H1_001N",
                           state = "Florida",
                           year = 2020,
                           geometry = TRUE
)

fl_city <- urban_areas(year = 2020) |> 
  filter(str_detect(NAME10, "FL"))
```

A [core-based statistical area](https://en.wikipedia.org/wiki/Core-based_statistical_area) is a generous definition of the area of interest, since it contains an urban nucleus (say, Houston) and also all neighboring communities that are tightly integrated with it. It is important to note that some of the neighboring communities share boundaries with the urban nucleus, but some are disconnected.

```{r}
jacksonville_cbsa <- core_based_statistical_areas(year = 2020) |>
  filter(str_detect(NAME, "Jacksonville, FL"))
```

A much stricter definition of the area of interest would result from using the city boundaries, which can be obtained using `urban_areas()`:

```{r}
jacksonville_city <- urban_areas(year = 2020) |> 
  filter(str_detect(NAME10, "Jacksonville, FL"))
```

Plot the location of Houston CBSA within Texas and its tracts:

```{r}
ggplot() +
  geom_sf(data = fl_tracts,
          fill = "lightgrey") +
  geom_sf(data = jacksonville_cbsa,
          color = "red",
          fill = NA) +
  geom_sf(data = fl_city,
          color = "blue",
          fill = NA) +
  theme_minimal()
```

Select all urban areas in Houston CBSA:

```{r}
jacksonville_urban <- st_filter(fl_city, jacksonville_cbsa)
```

Plot urban boundaries within the Phoenix CBSA:

```{r}
ggplot() +
  geom_sf(data = jacksonville_cbsa,
          fill = "lightgray") +
  geom_sf(data = jacksonville_urban,
          color = "blue",
          fill = NA) +
  theme_minimal()
```

How many named areas are there in this map?
```{r}
ggplot() +
  geom_sf(data = jacksonville_cbsa,
          fill = "lightgray") +
  geom_sf(data = jacksonville_urban,
          aes(fill = NAME10)) +
  theme_minimal()
```

Filter Houston:
```{r}
jacksonville_urban <- jacksonville_urban |>
  filter(NAME10 == "Jacksonville, FL")
```

Plot Houston urban:
```{r}
ggplot() +
  geom_sf(data = jacksonville_cbsa,
          fill = "lightgray") +
  geom_sf(data = jacksonville_urban,
          color = "blue",
          fill = NA) +
  theme_minimal()
```


Obtain the Houston tracts:
```{r}
jacksonville_tracts <- fl_tracts |>
  st_filter(jacksonville_urban, 
            .predicate = st_intersects)
```

Plot Phoenix's urban tracts within the CBSA:

```{r}
ggplot() +
  geom_sf(data = jacksonville_cbsa,
          fill = "lightgray") +
  geom_sf(data = jacksonville_tracts,
          color = "blue",
          fill = NA) +
  theme_minimal()
```

Save data objects:
```{r}
# Tracts
save(jacksonville_tracts, 
     file = glue::glue(here::here(),
                       "/data/jacksonville_tracts.rda"),
     compress = "bzip2")

# Urban boundaries
save(jacksonville_urban, 
     file = glue::glue(here::here(),
                       "/data/jacksonville_urban.rda"),
     compress = "bzip2")

# CBSA boundary
save(jacksonville_cbsa, 
     file = glue::glue(here::here(),
                       "/data/jacksonville_cbsa.rda"),
     compress = "bzip2")
```
