==> data-raw/00-Nashville-Tennessee-Basic-Data.Rmd <==
---
title: "Nashville Tennessee Basic Data"
---

# Nashville Tennessee Data

This will be the basic data pre-processing work for Nashville Tennessee. The other city notebooks follow the same steps but with different sets of CBSA data.

Estimated time to run: a few minutes

Load packages:

```{r}
library(ggplot2)
library(glue) # Interpreted String Literals
library(h3)
library(here) # A Simpler Way to Find Your Files
library(osmdata) # package for working with streets
library(sf) # Simple Features for R
library(tidycensus) # Load US Census Boundary and Attribute Data as 'tidyverse' and 'sf'-Ready Data Frames
library(tidyverse) # Easily Install and Load the 'Tidyverse'
library(tigris) # Load Census TIGER/Line Shapefiles
```

## Census tracts

Obtain the census tracts. Note that [{tidycensus}](https://walker-data.com/tidycensus/) function `get_decenial()` allows the selection of state and county, but not metro area or city. Check the example [here](https://walker-data.com/census-r/spatial-analysis-with-us-census-data.html#identifying-geometries-within-a-metropolitan-area) to see how to select tracts in a city.

```{r}
options(tigris_use_cache = TRUE)

tn_tracts <- get_decennial(geography = "tract",
                           variables = "H1_001N",
                           state = "Tennessee",
                           year = 2020,
                           geometry = TRUE
)

tn_city <- urban_areas(year = 2020) |> 
  filter(str_detect(NAME10, "TN"))
```

A [core-based statistical area](https://en.wikipedia.org/wiki/Core-based_statistical_area) is a generous definition of the area of interest, since it contains an urban nucleus (say, Houston) and also all neighboring communities that are tightly integrated with it. It is important to note that some of the neighboring communities share boundaries with the urban nucleus, but some are disconnected.

```{r}
nashville_cbsa <- core_based_statistical_areas(year = 2020) |>
  filter(str_detect(NAME, "Nashville"))
```

A much stricter definition of the area of interest would result from using the city boundaries, which can be obtained using `urban_areas()`:

```{r}
nashville_city <- urban_areas(year = 2020) |> 
  filter(str_detect(NAME10, "Nashville"))
```

Plot the location of Atlanta CBSA within Georgia and its tracts:

```{r}
ggplot() +
  geom_sf(data = tn_tracts,
          fill = "lightgrey") +
  geom_sf(data = nashville_cbsa,
          color = "red",
          fill = NA) +
  geom_sf(data = tn_city,
          color = "blue",
          fill = NA) +
  theme_minimal()
```

Select all urban areas in Atlanta CBSA:

```{r}
nashville_urban <- st_filter(tn_city, nashville_cbsa)
```

Plot urban boundaries within the Phoenix CBSA:

```{r}
ggplot() +
  geom_sf(data = nashville_cbsa,
          fill = "lightgray") +
  geom_sf(data = nashville_urban,
          color = "blue",
          fill = NA) +
  theme_minimal()
```

How many named areas are there in this map?
```{r}
ggplot() +
  geom_sf(data = nashville_cbsa,
          fill = "lightgray") +
  geom_sf(data = nashville_urban,
          aes(fill = NAME10)) +
  theme_minimal()
```

Filter Houston:
```{r}
 nashville_urban <- nashville_urban |>
  filter(NAME10 == "Nashville-Davidson, TN")
```

Plot Atlanta urban:
```{r}
ggplot() +
  geom_sf(data = nashville_cbsa,
          fill = "lightgray") +
  geom_sf(data = nashville_urban,
          color = "blue",
          fill = NA) +
  theme_minimal()
```


Obtain the tracts:
```{r}
nashville_tracts <- tn_tracts |>
  st_filter(nashville_urban, 
            .predicate = st_intersects)
```

Plot urban tracts within the CBSA:
```{r}
ggplot() +
  geom_sf(data = nashville_cbsa,
          fill = "lightgray") +
  geom_sf(data = nashville_tracts,
          color = "blue",
          fill = NA) +
  theme_minimal()
```

Save data objects:
```{r}
# Tracts
save(nashville_tracts, 
     file = glue::glue(here::here(),
                       "/data/nashville_tracts.rda"),
     compress = "bzip2")

# Urban boundaries
save(nashville_urban, 
     file = glue::glue(here::here(),
                       "/data/nashville_urban.rda"),
     compress = "bzip2")

# CBSA boundary
save(nashville_cbsa, 
     file = glue::glue(here::here(),
                       "/data/nashville_cbsa.rda"),
     compress = "bzip2")
```
