==> data-raw/00-Phoenix-Arizona-Basic-Data.Rmd <==
---
title: "Phoenix Arizona"
---

# Phoenix Arizona Data

This will be the basic data pre-processing work for Phoenix, Arizona. The other city notebooks follow the same steps but with different sets of CBSA data.

Estimated time to run: a few minutes

Load packages:

```{r}
library(ggplot2)
library(glue) # Interpreted String Literals
library(h3)
library(here) # A Simpler Way to Find Your Files
library(osmdata) # package for working with streets
library(sf) # Simple Features for R
library(tidycensus) # Load US Census Boundary and Attribute Data as 'tidyverse' and 'sf'-Ready Data Frames
library(tidyverse) # Easily Install and Load the 'Tidyverse'
library(tigris) # Load Census TIGER/Line Shapefiles
```

## Network data

Download OSM data for Arizona [here](https://download.geofabrik.de/north-america/us/arizona.html). It will be a proto-buffer file (pbf). Save under `data-raw/r5_graph/$CITY_NAME$`.

## Census tracts

Obtain the census tracts for Phoenix. Note that [{tidycensus}](https://walker-data.com/tidycensus/) function `get_decenial()` allows the selection of state and county, but not metro area or city. Check the example [here](https://walker-data.com/census-r/spatial-analysis-with-us-census-data.html#identifying-geometries-within-a-metropolitan-area) to see how to select tracts in a city.

```{r}
options(tigris_use_cache = TRUE)

az_tracts <- get_decennial(geography = "tract",
                           variables = "H1_001N",
                           state = "Arizona",
                           year = 2020,
                           geometry = TRUE
)

az_city <- urban_areas(year = 2020) |> 
  filter(str_detect(NAME10, "AZ"))
```

A [core-based statistical area](https://en.wikipedia.org/wiki/Core-based_statistical_area) is a generous definition of the area of interest, since it contains an urban nucleus (say, Phoenix) and also all neighboring communities that are tightly integrated with it. It is important to note that some of the neighboring communities share boundaries with the urban nucleus, but some are disconnected.

```{r}
phoenix_cbsa <- core_based_statistical_areas(year = 2020) |>
  filter(str_detect(NAME, "Phoenix"))
```

A much stricter definition of the area of interest would result from using the city boundaries, which can be obtained using `urban_areas()`:

```{r}
phoenix_city <- urban_areas(year = 2020) |> 
  filter(str_detect(NAME10, "Phoenix--Mesa"))
```

Plot the location of Phoenix CBSA within Arizona and its tracts:

```{r}
ggplot() +
  geom_sf(data = az_tracts,
          fill = "lightgrey") +
  geom_sf(data = phoenix_cbsa,
          color = "red",
          fill = NA) +
  geom_sf(data = az_city,
          color = "blue",
          fill = NA) +
  theme_minimal()
```

Select all urban areas in Phoenix CBSA:

```{r}
phoenix_urban <- st_filter(az_city, phoenix_cbsa)
```

Plot urban boundaries within the Phoenix CBSA:

```{r}
ggplot() +
  geom_sf(data = phoenix_cbsa,
          fill = "lightgray") +
  geom_sf(data = phoenix_urban,
          color = "blue",
          fill = NA) +
  theme_minimal()
```

How many named areas are there in this map?
```{r}
ggplot() +
  geom_sf(data = phoenix_cbsa,
          fill = "lightgray") +
  geom_sf(data = phoenix_urban,
          aes(fill = NAME10)) +
  theme_minimal()
```

Filter Phoenix-Mesa:
```{r}
phoenix_urban <- phoenix_urban |>
  filter(NAME10 == "Phoenix--Mesa, AZ")
```

Plot Phoenix urban:
```{r}
ggplot() +
  geom_sf(data = phoenix_cbsa,
          fill = "lightgray") +
  geom_sf(data = phoenix_urban,
          color = "blue",
          fill = NA) +
  theme_minimal()
```


Obtain the Phoenix tracts:
```{r}
phoenix_tracts <- az_tracts |>
  st_filter(phoenix_urban, 
            .predicate = st_intersects)
```

Plot Phoenix's urban tracts within the CBSA:

```{r}
ggplot() +
  geom_sf(data = phoenix_cbsa,
          fill = "lightgray") +
  geom_sf(data = phoenix_tracts,
          color = "blue",
          fill = NA) +
  theme_minimal()
```

Save data objects:
```{r}
# Tracts
save(phoenix_tracts, 
     file = glue::glue(here::here(),
                       "/data/phoenix_tracts.rda"),
     compress = "bzip2")

# Urban boundaries
save(phoenix_urban, 
     file = glue::glue(here::here(),
                       "/data/phoenix_urban.rda"),
     compress = "bzip2")

# CBSA boundary
save(phoenix_cbsa, 
     file = glue::glue(here::here(),
                       "/data/phoenix_cbsa.rda"),
     compress = "bzip2")
```
