==> data-raw/02-Houston-Census-Data.Rmd <==
---
title: "Houston Census Data"
---

# Setup

```{r load packages}
library(tidycensus)
library(tigris)
library(dplyr) # A Grammar of Data Manipulation
library(ggraph) # An Implementation of Grammar of Graphics for Graphs and Networks
library(ggplot2) # Create Elegant Data Visualisations Using the Grammar of Graphics
library(glue) # Interpreted String Literals
library(here) # A Simpler Way to Find Your Files
library(igraph) # Network Analysis and Visualization
library(sf) # Simple Features for R
library(sfnetworks) # Tidy Geospatial Networks
library(tidygraph) # A Tidy API for Graph Manipulation
library(tidyr) # Tidy Messy Data
library(r5r) # Rapid Realistic Routing with 'R5'
library(h3) # R Bindings for H3 # R Bindings for H3
library(purrr) # Functional Programming Tools
library(units) # Measurement Units for R Vectors
library(ggpubr)
```


```{r load packages}
# Walksheds
load(file = glue::glue(here::here(),
                       "/data/houston_walksheds.rda"))

# Tracts
load(file = glue::glue(here::here(),
                       "/data/houston_tracts.rda"))

# Urban boundaries
load(file = glue::glue(here::here(),
                       "/data/houston_urban.rda"))

# CBSA boundary
load(file = glue::glue(here::here(),
                       "/data/houston_cbsa.rda"))
```

```{r add areas of walksheds}
walksheds <- houston_walksheds |>
  mutate(area = as.numeric(st_area(geometry)))
# units of area are m^2
```

# Downloading Census Data

```{r}
# Median income in past 12 months: B06011_001 (2022), B19013_001 (2020)

income <- get_acs(
  state = "Texas",
  geography = "tract",
  variables = "B06011_001",
  geometry = TRUE,
  output = "wide",
  year = 2022)
# sort by county, doesn't always work:
# county = c("Harris County", "Fort Bend", "Brazoria", "Galveston", "Chambers", "Liberty", "Montgomery", "Waller", "San Jacinto"),
```

```{r}
white_p <- get_acs(
  state = "Texas",
  geography = "tract",
  variables = "B02001_002",
  geometry = TRUE,
  output = "wide",
  year = 2022)
```

```{r}
black_p <- get_acs(
  state = "Texas",
  county = c("Harris County", "Fort Bend", "Brazoria", "Galveston", "Chambers", "Liberty", "Montgomery", "Waller", "San Jacinto"),
  geography = "tract",
  variables = "B02001_003",
  geometry = TRUE,
  output = "wide",
  year = 2022)
```

```{r}
asian_p <- get_acs(
  state = "Texas",
  county = c("Harris County", "Fort Bend", "Brazoria", "Galveston", "Chambers", "Liberty", "Montgomery", "Waller", "San Jacinto"),
  geography = "tract",
  variables = "B02001_005",
  geometry = TRUE,
  output = "wide",
  year = 2022)
```

```{r}
american_indian_alaska_native_p <- get_acs(
  state = "Texas",
  county = c("Harris County", "Fort Bend", "Brazoria", "Galveston", "Chambers", "Liberty", "Montgomery", "Waller", "San Jacinto"),
  geography = "tract",
  variables = "B02001_004",
  geometry = TRUE,
  output = "wide",
  year = 2022)
```

```{r}
pacific_islander_p <- get_acs(
  state = "Texas",
  county = c("Harris County", "Fort Bend", "Brazoria", "Galveston", "Chambers", "Liberty", "Montgomery", "Waller", "San Jacinto"),
  geography = "tract",
  variables = "B02001_006",
  geometry = TRUE,
  output = "wide",
  year = 2022)
```

```{r}
other_p <- get_acs(
  state = "Texas",
  county = c("Harris County", "Fort Bend", "Brazoria", "Galveston", "Chambers", "Liberty", "Montgomery", "Waller", "San Jacinto"),
  geography = "tract",
  variables = "B02001_007",
  geometry = TRUE,
  output = "wide",
  year = 2022)
```

# Set Up Dataframes

```{r rename variable columns}
names(income)[3] <- "income"
names(white_p)[3] <- "white_pop"
names(black_p)[3] <- "black_pop"
names(asian_p)[3] <- "asian_pop"
names(american_indian_alaska_native_p)[3] <- "american_indian_pop"
names(pacific_islander_p)[3] <- "pacific_islander_pop"
names(other_p)[3] <- "other_pop"
```

```{r drop census tract data geometries}
walksheds <- st_drop_geometry(walksheds)
income <- st_drop_geometry(income)
white_p <- st_drop_geometry(white_p)
black_p <- st_drop_geometry(black_p)
asian_p <- st_drop_geometry(asian_p)
american_indian_alaska_native_p <- st_drop_geometry(american_indian_alaska_native_p)
pacific_islander_p <- st_drop_geometry(pacific_islander_p)
other_p <- st_drop_geometry(other_p)
```


```{r}
houston_tracts_df <- houston_tracts |>
  left_join(walksheds |> select(GEOID, area), by = "GEOID") |>
  left_join(income |> select(GEOID, income), by = "GEOID") |>
  left_join(white_p |> select(GEOID, white_pop), by = "GEOID") |>
  left_join(black_p |> select(GEOID, black_pop), by = "GEOID") |>
  left_join(asian_p |> select(GEOID, asian_pop), by = "GEOID") |>
  left_join(american_indian_alaska_native_p |> select(GEOID, american_indian_pop), by = "GEOID") |>
  left_join(pacific_islander_p |> select(GEOID, pacific_islander_pop), by = "GEOID") |>
  left_join(other_p |> select(GEOID, other_pop), by = "GEOID")
```

```{r}
houston_tracts_df <- houston_tracts_df |>
  mutate(total_pop = (white_pop + black_pop + asian_pop + pacific_islander_pop + american_indian_pop + other_pop))
```


```{r}
houston_tracts_df <- houston_tracts_df |>
  mutate(
    white_prop = white_pop/total_pop,
    black_prop = black_pop/total_pop,
    asian_prop = asian_pop/total_pop,
    american_indian_prop = american_indian_pop/total_pop,
    pacific_islander_prop = pacific_islander_pop/total_pop,
    other_prop = other_pop/total_pop)
# might only need white_prop and black_prop
```

# Save Dataframe

```{r}
save(houston_tracts_df, 
     file = glue::glue(here::here(),
                       "/data/houston_tracts_df.rda"),
     compress = "bzip2")
```

Or, if already saved, just load the file instead of re-running the previous code:

```{r}
load(file = glue::glue(here::here(),
                       "/data/houston_tracts_df.rda"))
```

# Maps

```{r}
# Make sure Census Data covers all of the tracts
ggplot(houston_tracts_df, aes(fill = area)) + 
  geom_sf() + 
  theme_void() +
  scale_fill_viridis_c(option = "turbo") +
  labs(fill = "Area of Walkshed
in Square Meters") +
  geom_sf(data = houston_tracts, fill = NA, color = "purple")
```

```{r}
ggplot(houston_tracts_df, aes(fill = black_prop)) + 
  geom_sf() +
  theme_void() +
  scale_fill_viridis_c(option = "turbo") +
  labs(fill = "Proportion of
Black Population")
```

```{r}
ggplot(houston_tracts_df, aes(fill = white_prop)) + 
  geom_sf() + 
  theme_void() +
  scale_fill_viridis_c(option = "turbo") +
labs(fill = "Proportion of
White Population")
```


# Graphs

The first two graphs here are test graphs, do not use those.

```{r, black prop versus walkshed area}
ggplot(houston_tracts_df, aes(x = black_prop, y = area)) +
geom_point(size = 0.75) +
  geom_smooth(method = "lm", se = FALSE, color = "grey") +
  labs(x = 'Proportion of Other Populations in Census Tract', y = 'Proportion of Food Stamps Used in Census Tract') +
  theme_classic()
# for help with lines of best fit: https://aosmith.rbind.io/2018/11/16/plot-fitted-lines/
```

```{r, white prop versus walkshed area}
ggplot(houston_tracts_df, aes(x = white_prop, y = area)) +
geom_point(size = 0.75) +
  geom_smooth(method = "lm", se = FALSE, color = "grey") +
  labs(x = 'Proportion of Other Populations in Census Tract', y = 'Proportion of Food Stamps Used in Census Tract') +
  stat_regline_equation(
    aes(label =  paste(..eq.label.., ..adj.rr.label.., sep = "~~~~", label.x.npc = "right", label.y.npc = "top"))) +
  theme_classic()
# for help with lines of best fit: https://aosmith.rbind.io/2018/11/16/plot-fitted-lines/
# for help with adj r^2 values: https://rpkgs.datanovia.com/ggpubr/reference/stat_regline_equation.html 
```

```{r}
ggscatter(houston_tracts_df, x = "black_prop", y = "area", add = "reg.line", shape = 19, size = 1, xlab = "Proportion of Black Population in Census Tract", ylab = "Area of Walkshed in Square Meters", font.label = c(5, "plain", "black")) +
  stat_cor(label.x = 0.75, label.y = 2000000) #label placement
```

```{r}
ggscatter(houston_tracts_df, x = "white_prop", y = "area", add = "reg.line", shape = 19, size = 1, xlab = "Proportion of White Population in Census Tract", ylab = "Area of Walkshed in Square Meters", font.label = c(5, "plain", "black")) +
  stat_cor(label.x = 0.75, label.y = 2000000) #label placement
```

```{r}
ggscatter(houston_tracts_df, x = "income", y = "area", add = "reg.line", shape = 19, size = 1, xlab = "Average Income in Census Tract (USD)", ylab = "Area of Walkshed in Square Meters", font.label = c(5, "plain", "black")) +
  stat_cor(label.x = 110000, label.y = 2000000) #label placement
```
